+++ lsp/server/src/validateGraphql.ts
@@ -15,9 +15,11 @@ import {
     isTheDiagnosticSuppressed
 } from './diagnostic/DiagnosticSettings';
 import { OversizedRecord as OversizedRequest } from './diagnostic/gql/over-sized-record';
+import { MisspelledUiapi } from './diagnostic/gql/misspelled-uiapi';
 
 const diagnosticProducers: DiagnosticProducer<ASTNode>[] = [
-    new OversizedRequest(),
+    new MisspelledUiapi(),
+    new OversizedRequest()
 ];
 
 /**
